requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
email-validator==2.0.0
Flask-CORS==4.0.0
Werkzeug==2.3.7
PyMySQL==1.1.0 